requirements.txt
pandas
matplotlib
seaborn
scikit-learn
Pillow
python-whois
tkhtmlview
reportlab
requests